CardContainer.show_trigger property
Some card containers should not show a "trigger" in the upper right
corner and some should. This adds a property that can be configured in
app.json.

Also removes the functionality from SetPreviewCard and
SupplementaryArticlesModule to show the trigger based on whether there
are more cards to see. Instead that is accomplished with a property
binding in CardContainer, and only happens if show_trigger is true.

https://phabricator.endlessm.com/T10591