Fix factory crash when all slots are missing
It can happen, e.g. on an Order module, that all slots are optional, and
none of them are filled with submodules, and therefore the "slots": {...}
object is missing from app.json. If that was the case, then the factory
would crash when trying to retrieve a submodule. This fixes the crash and
adds a test.

https://phabricator.endlessm.com/T10587