Refactor HighlightsModule to use SetPreviewCard
This is a fairly drastic refactor of HighlightsModule. Instead of packing
secondary arrangements, it now packs cards. Currently it only works with
SetPreviewCard, since that has a load_content() method.

This makes a few tweaks for SetPreviewCard to become equivalent to what
HighlightModule used to do; we add a separator, and replace the
CardContainer title with our own header-card-type (intended to be a
TextCard.)

The necessary changes are also made to buffet.json and the buffet
preset's CSS to make the travel app look as it did before.

The same refactor could also be done for ThematicModule, but I'll open a
separate issue for that.

https://phabricator.endlessm.com/T10587