Filter sets in highlights module
Previously the highlights module was hardcoded to display non-featured
sets as its support sets. Instead, implement that with a filter so that
it's configurable in app.json.

Change the Buffet preset so that it works as it did previously, and
change the News preset to display featured sets on the front page.

https://phabricator.endlessm.com/T10587