Filter items, not sets
It's items that we want to avoid displaying duplicates of in two
different containers. Change the FILTER_SETS action to FILTER_ITEMS,
change the code in highlights module that sends it to send items instead
of sets, and move the code that does the filtering from suggested
categories to set preview.

https://phabricator.endlessm.com/T10472