postCard: never underallocate
We should make sure that we don't give our shadow frame less space
then it needs to not cut off its children.

The fact that this is fixes a large number of graphical glitches
shows that the spacings we got (from invision I assume) don't always
match up with the size of the label, margins and paddings of the
shadow frame we are trying to allocate. Would be good to figure
our where and possibly tweak some css values.

Still never the write call to cut off text, so we should have this
fail safe in place