Send feature item from HierarchicalSetModule
Previously, we used the Thematic module for the section
page in the News App. Now, we switched to Hierarchical
module.

Therefore, remove the unnecessary code from the previous
module, move it and adapt it for Hierarchical module.

NOTE THAT: the same "section-page" in our preset is used
to implement two different pages of the design: section
page and sub-section page.

* We can cover the sub-section case because each card is
contained within the arrangement of the module. So we can
access the featured item.

* We can't cover the section case because each card is an
arrangement by itself so a) the module can't tell when
these cards are loaded, because they loaded asynchronously
in SetPreviewCard, and b) SetPreviewCard has no way to
tell if its the first, at least not without adding more
special glue for this case to the module and the card
making them coupled. So we can't access the featured item.

For the section case we only send the current set model as
the featured item because we have to set some repeatable
color anyway.

https://phabricator.endlessm.com/T10787