Add animations property to Window module
The design team requested that we disable the page
transition animations in the News App since it is
still too slow.

Add a new property to the Window module to set whether
we should change the GtkStack transition type or not.
The default type is GTK_STACK_TRANSITION_TYPE_NONE.

https://phabricator.endlessm.com/T10792