endless_buffet: fix theming bug
I'm pretty sure we were overriding all title to 22px by accident.
We want that for TextCards that title each section, but not for
the titles in our thumb cards, which have fonts set in a responsive
way in buffet_core.css.

This would end up bumping the font was up as the card sized down
and getting things cut off
https://phabricator.endlessm.com/T10691