Exclude sets from search results
Having sets showing up in search results is problematic in
apps based on preset A and B, because they can't handle it.

We removed sets in Mesh and Aisle interactions in a previous
patch, but we didn't remove it for the results displayed in
the search box popup.

NOTE THAT this patch removes sets from the search box popup
but ALSO removes sets from the results page in apps based on
the Buffet interaction for consistency.

https://phabricator.endlessm.com/T10449