Add more spacing in between top menu titles
Specs call for more spacing in between top menu titles.
In theory this spacing amount ought to be determined
by the screen resolution size. However, it would
break modularity for the SideBySide arrangement
to know about the screen resolution. So instead
we scale down the breakpoints in a kind of
hacky way so that we get the spacing we need.

https://phabricator.endlessm.com/T10795