thumbCard: don't stomp on grid allocations
If we want a custom allocation routine we need to write our own
container, not use a grid and stomp on the allocation it gives its
children. Gtk is not at all built to handle that case and we were
seeing strange rendering errors.

Still a lot of work to do on this widget, we are setting things
like visibility and align inside of allocate which we don't want.
But this is a start and should fix our rendering problems.

https://phabricator.endlessm.com/T10691