Separate disclaimer link from disclaimer window
For the Prensa Libre app, we were passing the disclaimer modal window as part
of the disclaimer text into the Mustache template. For a strange reason, this
was adding an extra, empty paragraph tag into the html, which was making the
alignment of the whole disclaimer section act weird!

We now have a separate element to inject the disclaimer modal window, which is
currently used only by the Prensa Libre app.

https://phabricator.endlessm.com/T11079