Move _check_more_content to background thread.
We were seeing a bug where the tab button on
the hamburger basement template was being drawn
on screen _before_ receiving its latest coordinates.
As a result it would sometimes be drawn in the wrong
position.

The problem was that the space container was firing
its notification that the all-visible property had
changed during size_allocate, and thus before
other widgets on the page had the chance to finish
completing their own size allocations. SlidingPanel,
for one, had not finished getting its new allocation
before it was told to show the its child widget (the
frame containing the TabButton). As such, it showed
that child widget in the 'old', and hence wrong, position.

What this commit does is tells the SetGroupModule widget
to, when it gets the all-visible notification, move its
check to the background thread. This ensures that
_all_ widgets get to complete their allocation before
the sliding panel is told to show the tab button.

- Martin & Rory

https://phabricator.endlessm.com/T11101