SetGroupModule emits MODULE_READY
Previously, it was not possible to use any other module than a
HighlightsModule on the front page of a buffet app, because the
BuffetInteraction required a module on its front page to send the
MODULE_READY action before it would start.

This causes the SetGroupModule to send MODULE_READY when it is ready to
display as well.

There are no tests for this, because MODULE_READY is one of the things I
would like to get rid of soonest in the refactors.

https://phabricator.endlessm.com/T11172