modules: add css names for gtk widget modules
We use style classes and css names really inconsistently. I'm itching
to fix that up. But in the interest of keeping the 3.20 change limited
in scope lets just give a css name that's the same as our gtype names.

We can't do this for windows right now, because our sdk css depends
on styling EosWindows. We have one css name per widget class, which
overrides instead of inheriting the parent css name. So if we gave
a css name of EknWindow we would lose EosWindow.

https://phabricator.endlessm.com/T11000