paperTemplate: extend bin
Frames will now whine if you stomp on their size allocate
routines (as they have internal widgets such as a label they need
to allocate). Better to subclass bin here, which is meant for a
custom size allocate.

Only downside is bin is not gadgetized and will not follow the box
model, if we want that, we could put a custom layout container
inside of the frame, grid, or box.
https://phabricator.endlessm.com/T11000