Allow mix and match selection of arrangements and cards in picard
Extend picard to allow user to select both an arrangement
and a card type with which to populate the page.

Both arrangement and card-type can now change within a
single run of the application, since the user specifies
them via combo boxes and not via the command line
as before.

This also adds our default CSS styling so the user can
see how the cards would actually look in an app.

https://phabricator.endlessm.com/T11297