Add .infobox_v2 class to reader2 scss
QA reported that the infobox shown in the right column of travel
app articles is placed too close to the disclaimer text.

What happened is that the article infobox is using .infobox_v2
class, but we only set rules for the .infobox class.

Both can be treated equally, so make sure we do that in reader2
scss. NOTE this patch also fixes the visual style for articles
that use .infobox_v2.

https://phabricator.endlessm.com/T11315