Define slots as part of Module.Class
Module.Class now looks for a Slots property which, like Properties and
Signals, will define the module's slots at construct time. Each key of
Slots is a slot name, and its value is (currently) an empty object. In
subsequent commits, this object will hold metadata about the slots.

This changes get_slot_names() to be a class method.

Slots from all implemented interfaces, and slots from the parent class,
will be pulled in. The slot information itself is stored as a frozen
object on TheModule.__slots__.

This converts all existing modules to use the new syntax to define their
slots. While doing so, it makes sure that the slot names in the
documentation of each module are current, and it gets rid of a few doubly
defined slots that were there.

https://phabricator.endlessm.com/T11180