encyclopediaWindow: remove font scaling
Font scaling is really damn broken at the moment. I think the fact
that our search results are in a scrolled window, and we have a
separate bunch of font overrides for composite means we won't
lose any resolution support.

If we really want the scaling back we can fix it up and bring it
back for encyclopedia. But I would be excited to move all our
knowledge apps off it.