Replace fake property overrides
In older versions of GJS there was no GObject.ParamSpec.override(). Now
there is, and so we use it instead of faking a property override on
classes that implement interfaces with properties.

This was masking a bug all along, where you have to provide getters and
setters for properties that you override on interfaces from C.

Annoyingly, it also makes Jasmine's objectContaining() not handle the
properties overridden from C, so we have to rewrite the history model
test somewhat.

https://phabricator.endlessm.com/T11180