Add an _has_extra_cards variable to track when to show "see more" button
We need to keep track of whether there are extra cards on this
module, so we can decide whether to show the "See More" button.

This is a bit awkward, since what we really are trying to do
is have modules depend on each other which we won't be able
to do until 2.7.

https://phabricator.endlessm.com/T11101