Check has_more_content variable to determine whether to show button
We should always check to see if we have more content to show
on the home page, even when returning from the all categories
page.

This is to account for the case where the window resolution
is small and you are on the categories page; you then maximize
the window, and return to the home page. In that case we want
to check again to see if we need to show the see more button.

https://phabricator.endlessm.com/T11101