Proof of concept for JSDoc3
This proof of concept explores what we would need to do to create a custom
plugin for JSDoc3, with the objective of improve the documentation generated by
default, by extending the handling of the abstract syntax tree (ast) to better
accomodate GObject syntax.

Our test plugin, `testplugin.js`, traverses the ast in search for new class
declarations, and for each class declaration, it prints a little summary of its
information, properties and comments.

The output of a test run over the `js/app/widgets` directory is included in
`testplugin.out`.

The `jsdoc3conf.json` is a minimal configuration file to run our project.

https://phabricator.endlessm.com/T11202