WIP: unify name tracking
Make a first pass over the content of the app json in order
to map all the modules descriptions with their names.

If we don't do this, we will need to traverse the app json
every time we want to find a description that is not at the
top level of the app json.

https://phabricator.endlessm.com/T11178