WIP: support reference to other modules
This approach is based on having the 'reference' key in slots
inside the class definition, and let the factory infer whether
or not a module should be 'reference-able'.

The factory checks each time a named module it's mentioned during
it's first pass on the app json data. If a slot is referencing a
name module and the slot definition says its a reference, then the
factory will 'mark' that name so it can get hold of the instance
reference the first time the module is created.

This way there the factory can pass the reference to any other
module asking for a reference to the that module.

https://phabricator.endlessm.com/T11178