Remove EncyclopediaCoverTemplate
This replaces the EncyclopediaCoverTemplate with a TwoPieceTemplate.
The EncyclopediaCoverTemplate's gimmick was to have a grid inside a
frame. The TwoPieceTemplate is just a grid, so we take the frame's
background and put it on the window instead. This is more consistent with
what the other apps do.

We also replace the search box's min-width-chars, which was relying on
some odd sizing detail, with CSS min-width.

https://phabricator.endlessm.com/T11474