Autobahn, the app.json preprocessor
This adds a preprocessor utility and a simple test framework for an
app.json preprocessor in YAML, with two examples.

One is a basic example just to make sure that the correct app.json format
comes out. The other is a port of the existing encyclopedia.json to the
Autobahn format. It illustrates how to inherit from existing YAML objects
with <<.

https://phabricator.endlessm.com/T11516