references: support ids and references slots
Our knowledge apps are defined in a tree structure inside the
app json file. Each node of that tree describes a module, it's
properties and it's sub-modules.

Currently, modules can access their sub-modules instances but
they can't access other modules instances in the tree so we
rely on properties names and messages through the dispatcher,
which requires adding custom logic and custom messages in many
cases.

This patch introduces a mechanism for allowing modules to refer
and access to other modules instances in the tree.

Modules can use an "id" key in the json file, so other modules
can refer to them. To reference modules, other modules can use
"references" slots where each slot contains the id of the module.

These ids must be unique, therefore references are limited to
modules that are created once, e.g. not slots containing cards.

https://phabricator.endlessm.com/T11178