Remove compatibility with EKNv1 apps
In the brave new xdg-app world, we no longer have to worry about
backwards compatibility support for on-disk data, so we can effectively
remove all DomainV1 code as it'll never be used.

At somep oint there should be a much larger scale removing of
compatibility paths, but for now I'll just remove the code paths which
my code affects.

T11489