Query big enough batch size for sort and filter
Previously, we queried only 3 articles if the arrangement could display
3. This caused a bug with sorting or filtering since the order and filter
are not reflected in the Xapian database query. We would query a few
articles in database order, then sort / filter those.

For example, if we are sorting by a fictitious integer and querying for
an arrangement that displays only 3 cards:

  DB      Query 3   Display to user
  ==      =======   ===============
  5 }     5         1
  1 }---> 1         5
  9 }     9         9
  2
  7
  4
  3
  ...

When really what should be displayed to the user is 1, 2, 3.

This is a workaround until we can have the functionality to actually
reflect our sorting and filtering in Xapian Bridge queries.

https://phabricator.endlessm.com/T11567