Don't create cards beyond arrangement max
For arrangements that advertise having a max on the number of cards, we
don't create a card widget when a model is added to the arrangement that
wouldn't otherwise be shown.

If there is no ordering, then card widgets are simply not created if the
number of models is above the max.

If there is an ordering, then the card widget is only created if its
position indicates that it would be shown on screen; that is, under the
max.

https://phabricator.endlessm.com/T11567