Don't consider tags that aren't sets
This (mostly) works around a problem with some recent shards in the
Prensa Libre app, but I think it's a good piece of code to have anyway.
When there are tags that don't map to sets, then HierarchicalSet should
not consider them for its decision on whether to show an article or not.

https://phabricator.endlessm.com/T11584

Conflicts:
	js/app/modules/hierarchicalSetModule.js
	tests/js/app/modules/testHierarchicalSetModule.js