css: remove odd theming image on image after text card
Can't see the responsive margin breakages referred to in the comment
so just going to go ahead and remove this.

They are showing up as little spinners, because themableImage uses
gtk_render_activity. Was setting the icon to none to try and avoid
that, but apparently that broke with the gkt upgrade. Will have to
investigate.

For now lets remove the broken theming.

https://phabricator.endlessm.com/T11336