Remove named modules from compat JSON
This is the first step in getting rid of named modules: remove them from
our compat JSON and make sure everything still works.

This takes all the named modules and expands them throughout the JSON
files so that in the end there is only one named module, "interaction",
at the root of the module tree.

We have to change some code in compat.js which was looking certain
modules up by name, in order to deal with legacy app.jsons. Now
everything is looked up by type.

https://phabricator.endlessm.com/T11516