Discontinue named modules
This removes support for named modules in the factory. Now the only key
allowed under the "modules" key in app.json is the "interaction" key.
Removes the create_named_module() API from the factory, which is good,
since nothing uses it anymore except for InteractionLoader; and with that
we've eliminated one of the last vestiges of the previous non-modular
system.

The tests still extensively use create_named_module(), which remains a
method on MockFactory. I'll solve this in a subsequent commit.

https://phabricator.endlessm.com/T11516