Rename key for root of module tree
Since there is now only one root of the module tree, it doesn't make
sense to have it be a sub-key of a "modules" key in app.json. This
removes

    "modules": {
      "interaction": {

and replaces it with

    "root": {

Also updates Autobahn to generate the new structure.

Because of the mass dedent, it's best to view this diff while ignoring
whitespace changes.

https://phabricator.endlessm.com/T11516