Fix errors in make implicit rules
It seems that when an error occurs in an implicit rule (like .yaml.json:)
make will just keep going if the output file is created. To make sure we
error out, we remove the build product if an error occurs in implicit
rules, for both Autobahn (in its code) and SCSS (which we don't control,
so in the rule.)

Also improves the Autobahn error message a bit by removing irrelevant
stack trace.

https://phabricator.endlessm.com/T11516