ModuleFactory can handle arrays of submodules
We want the ModuleFactory to handle the creation of several submodules for a
single slot, as long as the contents of said slot are encoded in the app.json as
an array. Such array may contain an arbitrary number of elements.

The new method in the ModuleFactory will then return an array of submodules, as
opposed to a single submodule.

We also modify autobahn to account for these changes, making it aware of these
arrays and allowing it to traverse them recursively.

https://phabricator.endlessm.com/T11444