Added BoxLayout
The BoxLayout is a simple layout component that that places an arbitrary number
components next to each other (if the orientation is horizontal), or below each
other (when the orientation is vertical.

This is the first module that can have a variable number of submodules, and we
manage that by using an array of submodules in the app.json.

Now that we have a BoxLayout module, we decomission the following simple layouts
that were used in the past:
 - TwoPieceTemplate
 - BannerTemplate
 - DividedBannerTemplate

To achieve visual parity, we add a Separator module, which wraps a Gtk.Separator
and adds it in between submodules of the BoxLayout as done previously by the
BannerTemplate.

https://phabricator.endlessm.com/T11444