Support static and dynamic titles on ContentGroups
Titles for content groups can either be set statically,
in the app.json file, or can be dynamic. If dynamic,
the title is determined by the collection, which then
sets a title property which is bound to the title
property of the content group.

Static titles always override dynamic ones.

https://phabricator.endlessm.com/T11465