Add error check and informative error message in ModuleFactory
Previously, if you had a situation where you forgot to define
a slot name in a module class (a not uncommon mistake), you
would get a cryptic error message from the factory about
how 'in' was not an operator on undefined.

This instead explictly checks to see if the slot you are
referencing was defined in the module class, and, if not,
throws an error with a suitable message.

https://phabricator.endlessm.com/T11465