default theme: initial version
This is the initial version of the default theme, it is
based on Joana's notes (colors palette and fonts) and it
is heavily based on news app.

Add a partial scss file for every* module. Each of these
files provides the styling properties of each module, so
we can start definiting these syles more module-centric.

Add a default scss file which imports all these partial
files, and acts as the default theme.

*Note that this initial version only covers the modules
used in News and Buffet templates.

https://phabricator.endlessm.com/T11280