articleHTMLRenderer: Replace server-template-v1 with isServerTemplated
Will objected to the use of the 'source' field to mark an article as not
requiring server templating, so let's use a separate field
here. Additionally, since I wrote this code, we have settled on the idea
of targeting an SDK version, so we don't need a "v1" marker in the
data itself, since the shard itself is for a specific SDK version.