Ensure natural >= minimal in responsive margins
I observed it happen, while dragging the size of the window, that the
responsive margins module came up with a natural size that was smaller
than the minimal size. This is probably happening because of our double
size allocations. It should be properly fixed with @media queries, but in
the meantime, ensure that natural is at least as large as minimal.

https://phabricator.endlessm.com/T11736