Split Card.Deck out of Card.Post
We originally thought each card module was going to have its own
appearance when showing a set. In practice this turned out not to be
true. So, we can simplify Card.Post by splitting out the "deck of cards"
drawing logic into a Card.Deck module.

Card.Deck is a copy of Card.Post but with the inner content grid removed,
since it does not display a context label.

We also move the margin on Card.Deck into CSS, and change the drawing
routines that draw the sleeve and the arrow to respect CSS margin.

https://phabricator.endlessm.com/T11736