Use min-height for Card.Post and Card.Deck frames
This gets rid of more code (and complicated
Gtk.Overlay.get_child_position code at that) in favour of CSS rules.

Since according to the CSS spec min-height does not include padding, we
have to get a bit tricky with the interplay between Card.Post's padding
and min-height on its .content-frame.

https://phabricator.endlessm.com/T11736