Use module trees in tests
The mock factory was getting out of sync with the module factory. This
came to be a problem as we are adding lots of functionality to the module
factory and don't want to have to add it twice.

Tests now pass an app.json tree to MockFactory's constructor, and use
create_module_tree() to create it just like the real code does. The
difference is that instead of string values for the "type" key, you must
pass the real constructor objects that you have imported yourself. The
warehouse is bypassed. This allows you to use the modules from Minimal or
MockWidgets that the warehouse doesn't know about.

This gets rid of all created named mocks in the mock factory (since we
have also gotten rid of named modules in the real factory). Now, if you
want access to created children, you use get_created() or
get_last_created(). These functions take a string, which is a dotted path
of slots. (e.g.: "content.arrangement.card-type".) No "root." is prefixed
and no numbers are added for multi-slots.

https://phabricator.endlessm.com/T11746