Remove all those "it can be constructed" tests
This was my fault for introducing them, but they are inconsistently
implemented and also useless because all the other tests in the same file
cover the same code path.

I have kept only one, in ImageCoverFrame, because there are no other
tests in that file, so it is actually exercising a code path that
wouldn't otherwise get exercised.

https://phabricator.endlessm.com/T11746