Use class names consistently
We use the new dotted module names for all Name parameters, as well as
for the names of all those modules' test suites, so that it's clear what
module a test failure comes from.

At the same time, make sure all class Name strings are equal to the class
name in code.

We also organize the .ui files to correspond to the new naming scheme.

https://phabricator.endlessm.com/T11736