property bindings: add support in module factory
In some cases We want one module's property value to depend on
another module's property value. So far we have done this in a
limited way by assuming submodule types and the existance of
submodules properties or methods, which is not very modular.

This patch introduces a more general approach which allows to
define property bindings between modules in the app json. This
is basically a wrapper for g_object_bind_property, and is based
on module modules references.

A property binding is defined by:

* source-id, the id of the module that owns the property.
* property, the name of the property owned by that module.
* invert, whether the property value should be inverted.

A YAML example:

  type: Card
  properties:
    expand: true
    visible:
      binding:
        source-id: a-module-id
        property: visible
        invert: true

https://phabricator.endlessm.com/T11337