testEngine: Fix subtly wrong code in test
This test is actually subtly broken. Since the arity of the function
determines whether the test was asynchronous or synchronous, the test
actually was treated as synchronous and passed (with 0 assertions). The
only side effect was an errant traceback later in the code.

This was extremely difficult to track down.

The test is otherwise, completely broken, but now it runs correctly, at
least.