testEngine: Fix the test
There were several major issues wrong with the test here, preventing it
from working correctly.

We forgot to update it for the s/domain/app_id/g replacement, our spies
were not asynchronous, our callbacks had the wrong arity, and we were
calling the wrong function on the resulting task. Yikes.

Fix it all, fix the name of the test (!!), and remove extra leftover
dead code while we're at it, too.