card: update highlight after style context changes
This should allow us to pick up on cases when we want new styling
to apply to the highlighted label, because a style class was added
for example.

We need this now that some css classes are added automatically in
knowledge.js after instance _init, so we won't have the full selector
hierarchy ready when calling set_label_with_highlight.

Very unsure about this one, so sorry if I'm causing future bugs :)

Ideally the highlight string could be its own css node, capable of
notifying whenever its style properties change and updating its
own markup.

We could also consider figuring our how to rework the knowledge.js
auto class name code to apply itself as soon as this.parent is called
in the widget constructor.
https://phabricator.endlessm.com/T10053