window/app: restore focus after stack transition
My hacky stomp on all other drawing during a stack transition
broke a thing :astonished:. We will unfocus any child of the stack
the has focus when we set it insensitive, and not restore that focus
when things are made sensitive again.

Setting everything insensitive is still a total hack, but one we
want till we have the gl renderer I think. We really don't want
to be drawing our search entry again during the stack transition
because of a blinking cursor.

So for now, just restore focus after setting thing sensitive in
window.
https://phabricator.endlessm.com/T11266