Remove .*-page-{a,b} style classes
Didn't think we'd need this refactor so soon, but here it is. In
preparation for the pager, we remove the .home-page-a, .section-page-b,
etc. style classes, and replace them with .home-page, .section-page, etc.

We also add .preset-a and .preset-b classes to the window in order to be
able to select Preset A and B apps in the CSS files. However, these will
go away soon too, as part of https://phabricator.endlessm.com/T10097.

In order to make future splitting out of Preset A and B theming easier,
I grouped all the .preset-a and .preset-b rules together at the bottom of
mesh.scss.