Layouts render a background
Previously these widgets didn't render a background, and so
background-image in CSS would not work. This changes Layout.Sidebar and
Layout.SideMenu to render a background, and will therefore work with
pagers putting a background image on them.

This functionality was already present in several places, so it's
extracted out into a Utils function.

https://phabricator.endlessm.com/T11992