layout/sideMenu: clean up menu padding/margin
We had a lot of margin that were basically trying to imitate padding
on the menu itself, way easier and less error prone to just switch
over. Probably holdover from pre 3.20 where a lot of this was margin
set in code.

Also allows us to get rid of the questions sidebar class applied
to the module in our sidebar menu. Before we had a menu class applied
to what anyone looking at the widget would actually call the sidebar,
and a sidebar class applied to the arrangement contents inside it.
https://phabricator.endlessm.com/T10053