theme: move in article search styling to common
Not touching the default theme now, as I'm not sure what the plan
is for moving elements that are not in our travel and news apps to
the default theme.

But there is not reason for us to carry different versions of the
in article search in different places. This styling is for the most
part ad-hoc anyway.

Tried to take the best of all worlds and consolidate in _common.scss
https://phabricator.endlessm.com/T10053