Decouple page switching from window modules
This moves page switching out of Window.App and Window.Encyclopedia, into
a pair of Pager modules.

Pager.Simple has six slots for fixed pages, and switches between them on
the appropriate dispatcher actions. (Will be changed to respond to global
state.) All slots are optional except the home page.

There is one new CSS property, -EknPager_Simple-transitions-style, which
can be one of the strings 'none' (turn off page animations altogether as
we do in the travel and Prensa Libre apps), 'slide-all' (what happened on
Preset A), 'slide-center' (what happened on Preset B), and 'splash' (what
happened on Encyclopedia.) This style property is somewhat of a hack;
ideally we would be able to set the transitions through CSS.

Pager.ParallaxBackground is the same as Pager.Simple except it does the
required calculations to allow a parallax background transition.

Preset B currently uses Pager.ParallaxBackground since that's what it did
before; but I'm not sure if that was by design.

https://phabricator.endlessm.com/T11992