Only fade in cards that are newly created
We now have a setup where we unpack and repack
all cards in an arrangement whenever the set of models
backing that arrangement changes. Consequently, an card
whose fade-in parameter was set to true would unpack
itself and then fade itself in. This looks weird, so
instead let's only fade in cards which are new to the
arrangement.

https://phabricator.endlessm.com/T11233