Arrangement.List supports max-rows
In the case of this arrangement, max-rows is
equivalent to max-cards since there is only
ever one column. At some point we need to consolidate
the max-rows properties. Design has a plan to
be able to set max-rows, min-rows, num-cards-per-row etc.

That will all happen once we have the generic grid
arrangement that will replace many arrangements we
have now. This is the least invasive change to
allow for the port of highlights module to use
selections and content group.

https://phabricator.endlessm.com/T11233