historyStore: remove dependence on ekn-history-model
We can replicate everything we need from the history model with
just a few extra lines of javascript in our history store.

There's absolutely nothing wrong with the history model api
but I think its outlived its usefulness. Simpler to keep all of our
history machinery to javascript and be able to adapt it quicker to
the new architecture we are aiming for in with the history store.

https://phabricator.endlessm.com/T11266