pages: Refactor page names and rename section page
The page name constants were common to the controllers, so factor them
out into a separate enum, that can also be used in tests.

Section was an old word for the set page, we even called a page in
the travel app the all-sets page. This will hopefully make things
a bit clearer.

(Matt W & Philip C)

https://phabricator.endlessm.com/T11266