Remove extra provider in Pager.ParallaxBackground
We still need to do disk I/O on the background image to find out its
height/width ratio, unfortunately. But if we add CSS classes based on
whether the pager's ratio is greater or less than the image's, then we
can at least get rid of the extra CSS provider, which should improve
performance a bit.

https://phabricator.endlessm.com/T11992