application: dispatch app activating events
For the most part we are going to be removing actions from the
dispatcher, but what was previously the launcher methods are actually
an appropriate place to add some in.

They are user generated events (even if they weren't from this process)
that should update the application's history state.

This also removes the PRESENT_WINDOW action, since the above change would
otherwise have broken it. We now keep track of present timestamps in the
history items, instead of shoving them through the dispatcher.

(Matt W & Philip C)

https://phabricator.endlessm.com/T11266