engine: Add a EKN GVfs scheme handler
This is a major code refactor that lets us merge several important code
paths, and allows us to use ekn resources in e.g. GtkImage
naturally.

The GFile that's here is completely barebones right now, and doesn't do
much, but it's enough to get the Prensa Libre app running. I haven't
tested it if works on Wikipedia apps yet...