actions: remove page show actions
We remove the SHOW_*_PAGE actions and instead rely on history store
change notifications.

In a few places, the controllers would wait for something to load and
only then dispatch a page show action. Now we don't have this, since the
history store change is processed immediately. Instead, we use
make_ready() to load whatever needs to be loaded, and change the pager so
that it only turns the page after the new page has declared itself ready.

(Matt W & Philip C)

https://phabricator.endlessm.com/T11266