Fix inline translated strings' POT comments
Previously putting a translated string inline in a shortdef, e.g.
"shortdef: 'Module(prop: !translate message)'", would result in an
incorrect comment in the POT file telling where the string came from,
e.g. "#: <unicode string>".

We still can't get a line number for these strings, since the inline
definition is parsed as a separate mini-YAML document, but we can at
least add the correct file name so the comment points you in the right
direction if you're looking for the string.

https://phabricator.endlessm.com/T10450