Full stack traces in AsyncTask
I thought we could discard all but the innermost invocation of
catch_callback_errors() in AsyncTask, but it turns out that important
information is in those other stack traces as well. This makes stack
traces longer and therefore less readable, but nonetheless gives you all
the information needed for debugging.

https://phabricator.endlessm.com/T11266