Remove SHOW_ARTICLE action
This action was previously used to cause the article stack to transition,
and to give the lightbox its array of resources. For the lightbox, we can
simply move the affected code to a history state change. For the article
stack, we can move some code out of the controllers and stop passing
around a loading animation type enum.

https://phabricator.endlessm.com/T11266