Rework MediaLightbox to use history store
This adds a 'media-model' property to the history items, indicating a
content object model that should be shown in a lightbox on top of the
current page.

When the current history item has a non-null 'media-model' property, the
lightbox will open and display that media object. When the current
history item's 'media-model' is null, the lightbox closes.

This allows getting rid of the SHOW_MEDIA and HIDE_MEDIA dispatcher
actions.

It also requires adding another dispatcher action, LIGHTBOX_CLOSED, which
is initiated by the user clicking on the X button or outside of the
lightbox. The history stores handle it and change to a state with a null
'media-model'.

The user-noticeable change in this commit is that from the lightbox,
clicking the history back button will now close the lightbox and bring
you to the page you were on when you opened the lightbox, instead of
going back to the previous page entirely.

However, navigating through the list of images in the lightbox is not
controlled by the history state, so if you click an image ("Image 1"),
click the lightbox arrow to go forward to the next image ("Image 2"),
clicking the history back button will still close the lightbox.
Subsequently clicking the history forward button will bring you to Image
1, not Image 2, and the history forward button will be disabled at that
point. This could be addressed in a follow-up issue, but it's arguable
that the current behaviour is correct.

https://phabricator.endlessm.com/T11266