Close lightbox when search focused
There was code for this functionality, but it had become broken along the
way, as the window stopped emitting the search-focused signal.

Since focusing the search box is user input to the UI, we create a
dispatcher action for it, and the history stores handle it and dispatch
the HIDE_MEDIA action, which in a following commit will be turned into a
state change.